[Fix] #22: JWT Filter 및 토큰 추출 오류 수정

+++ src/main/java/com/example/off/domain/member/controller/MemberController.java
@@ -58,7 +58,7 @@ public class MemberController {
     }
 
     private Long getMemberId (HttpServletRequest req){
-        Object memberId = req.getAttribute("memberId");
+        Object memberId = (Long) req.getAttribute("memberId");
         if (!(memberId instanceof Long id)) { //변수 선언 동시에 처리
             throw new OffException(ResponseCode.INVALID_TOKEN);
         }
